Create different module for each tab and content div
