+++ src/app/components/header.js
@@ -1,7 +1,8 @@
+const mainDiv = document.getElementById('content');
 const header = document.createElement('header');
 const defaultText = 'Get the best from food';
 header.setAttribute('id', 'display-content');
 header.textContent = defaultText;
-document.body.appendChild(header);
+mainDiv.appendChild(header);
 
 export default defaultText;
