+++ src/template.html
@@ -4,8 +4,9 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
+  <title>Restaurant</title>
 </head>
 <body>
+  <div id="content"></div>
 </body>
 </html>
